requirements.txt
websocket-client
requests
psutil
pyyaml